refactor(hub): split hub-ipc into domain modules behind a facade
