+++ src/main/hub/hub-ipc.ts
@@ -1,1472 +1,40 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
-// IPC handler for Hub upload operations
+// IPC handler for Hub upload operations.
+//
+// This file is the facade: it owns no logic of its own and re-exports
+// the full public surface from the sibling modules in this directory:
+//
+//   hub-ipc-shared.ts             — input validators + small pure helpers
+//   hub-ipc-token.ts              — JWT cache + auth-retry wrapper
+//   hub-ipc-analytics-prepare.ts  — shared analytics-export assembly
+//   hub-ipc-posts.ts              — keymap post CRUD + auth-me handlers
+//   hub-ipc-favorite.ts           — favorite (feature) post handlers
+//   hub-ipc-analytics.ts          — analytics post upload/update/preview
+//   hub-ipc-private.ts            — private (unlisted) upload handlers
+//   hub-ipc-packs.ts              — i18n + theme pack handlers
+//   hub-ipc-key-labels.ts         — Key Label Hub handlers
+//
+// New Hub IPC logic belongs in the sibling module whose responsibility
+// it extends — not here. External consumers (sync-ipc.ts, main/index.ts,
+// and every test file's mock of this facade path) must keep importing
+// this facade path, never a submodule directly. See
+// .claude/tasks/backlog/Task-split-hub-ipc.md and
+// .claude/rules/file-splitting.md for the split rationale.
 
-import { secureHandle } from '../ipc-guard'
-import { IpcChannels } from '../../shared/ipc/channels'
-import { HUB_ERROR_DISPLAY_NAME_CONFLICT, HUB_ERROR_ACCOUNT_DEACTIVATED, HUB_ERROR_RATE_LIMITED } from '../../shared/types/hub'
-import type {
-  HubUploadPostParams, HubUpdatePostParams, HubPatchPostParams, HubUploadResult, HubDeleteResult,
-  HubFetchMyPostsResult, HubFetchMyKeyboardPostsResult, HubUserResult, HubFetchMyPostsParams,
-  HubUploadFavoritePostParams, HubUpdateFavoritePostParams,
-  HubUploadAnalyticsPostParams, HubUpdateAnalyticsPostParams, HubPreviewAnalyticsPostParams,
-  HubAnalyticsPreview, HubAnalyticsFilters, HubAnalyticsCategoryId,
-  HubPrivateUploadResult, HubPrivateKind,
-  HubPrivateUploadPostParams, HubPrivateUploadFavoritePostParams, HubPrivateUploadAnalyticsPostParams,
-} from '../../shared/types/hub'
-import { getIdToken } from '../sync/google-auth'
-import { Hub401Error, Hub403Error, Hub404Error, Hub409Error, Hub429Error, authenticateWithHub, uploadPostToHub, updatePostOnHub, patchPostOnHub, deletePostFromHub, fetchMyPosts, fetchMyPostsByKeyboard, fetchAuthMe, patchAuthMe, getHubOrigin, uploadFeaturePostToHub, updateFeaturePostOnHub, uploadAnalyticsPostToHub, updateAnalyticsPostOnHub, uploadPrivatePostToHub, uploadPrivateFeaturePostToHub, uploadPrivateAnalyticsPostToHub, deletePrivatePostFromHub } from './hub-client'
-import {
-  buildAnalyticsExport,
-  estimateAnalyticsExportSizeBytes,
-  sanitizeFingerOverrides,
-  validateAnalyticsExport,
-  type BuildAnalyticsExportInput,
-  type DeviceScope,
-} from './hub-analytics'
-import {
-  uploadI18nPostToHub,
-  updateI18nPostOnHub,
-  deleteI18nPostFromHub,
-  fetchI18nPostList,
-  downloadI18nPostBody,
-  validateI18nExport,
-  fetchI18nPackTimestamps,
-  type HubI18nListParams,
-  type HubI18nListResponse,
-} from './hub-i18n'
-import { setHubPostId as setI18nPackHubPostId } from '../i18n-pack-store'
-import {
-  uploadThemePostToHub,
-  updateThemePostOnHub,
-  deleteThemePostFromHub,
-  fetchThemePostList,
-  downloadThemePostBody,
-  fetchThemePackTimestamps,
-} from './hub-theme'
-import { setHubPostId as setThemePackHubPostId } from '../theme-pack-store'
-import { validateThemePack } from '../../shared/theme/validate'
-import type {
-  HubUploadI18nPostParams,
-  HubUpdateI18nPostParams,
-  HubI18nExportV1,
-  HubI18nPackTimestamp,
-  HubI18nPackTimestampsResponse,
-  HubUploadThemePostParams,
-  HubUpdateThemePostParams,
-  HubThemePackBody,
-  HubThemePackTimestamp,
-  HubThemePackTimestampsResponse,
-  HubThemeListParams,
-  HubThemeListResponse,
-} from '../../shared/types/hub'
-import { HUB_I18N_PACK_TIMESTAMPS_BATCH_LIMIT, HUB_THEME_PACK_TIMESTAMPS_BATCH_LIMIT } from '../../shared/types/hub'
-import { readAnalyzeFilterEntry, setAnalyzeFilterHubPostId } from '../analyze-filter-store'
-import { getKeymapSnapshotForRange } from '../typing-analytics/keymap-snapshots'
-import { getMachineHash } from '../typing-analytics/machine-hash'
-import type { TypingKeymapSnapshot } from '../../shared/types/typing-analytics'
-import type { LayoutComparisonMetric } from '../../shared/types/typing-analytics'
-import type { KleKey } from '../../shared/kle/types'
-import type { HubAuthResult, HubUploadFiles } from './hub-client'
-import { fetchKeyLabelList, fetchKeyLabelDetail, fetchKeyLabelTimestamps, downloadKeyLabel, uploadKeyLabel, updateKeyLabel, deleteKeyLabel } from './hub-key-labels'
-import type { HubKeyLabelInput } from './hub-key-labels'
-import type { HubKeyLabelItem, HubKeyLabelListResponse, HubKeyLabelListParams, HubKeyLabelTimestamp, HubKeyLabelTimestampsResponse } from '../../shared/types/hub-key-label'
-import { HUB_ERROR_KEY_LABEL_DUPLICATE, HUB_KEY_LABEL_TIMESTAMPS_BATCH_LIMIT } from '../../shared/types/hub-key-label'
-import { getRecord, saveRecord, setHubPostId } from '../key-label-store'
-import type { SaveRecordInput } from '../key-label-store'
-import type { KeyLabelMeta, KeyLabelRecord, KeyLabelStoreResult } from '../../shared/types/key-label-store'
-import { isValidFavoriteType, isValidHubVialProtocol, FAV_TYPE_TO_EXPORT_KEY, serializeFavData, buildFavExportFile } from '../../shared/favorite-data'
-import { serialize as serializeKeycode } from '../../shared/keycodes/keycodes'
-import { withSerializeProtocol } from '../../shared/keycodes/with-protocol'
-import type { FavoriteType, FavoriteIndex } from '../../shared/types/favorite-store'
-import { readFile } from 'node:fs/promises'
-import { join } from 'node:path'
-import { app } from 'electron'
-
-const AUTH_ERROR = 'Not authenticated with Google. Please sign in again.'
-const POST_ID_RE = /^[a-zA-Z0-9_-]+$/
-const DISPLAY_NAME_MAX_LENGTH = 50
-
-function validatePostId(postId: string): void {
-  if (!postId || !POST_ID_RE.test(postId)) {
-    throw new Error('Invalid post ID')
-  }
-}
-
-function validateDisplayName(displayName: unknown): string {
-  if (displayName == null || typeof displayName !== 'string') throw new Error('Display name must not be empty')
-  const trimmed = displayName.trim()
-  if (trimmed.length === 0) throw new Error('Display name must not be empty')
-  if (trimmed.length > DISPLAY_NAME_MAX_LENGTH) throw new Error('Display name too long')
-  return trimmed
-}
-
-const KEYBOARD_NAME_MAX_LENGTH = 100
-
-function validateKeyboardName(name: unknown): string {
-  if (typeof name !== 'string' || name.trim().length === 0) throw new Error('Missing keyboard name')
-  const trimmed = name.trim()
-  if (trimmed.length > KEYBOARD_NAME_MAX_LENGTH) throw new Error('Keyboard name too long')
-  return trimmed
-}
-
-const TITLE_MAX_LENGTH = 200
-
-function validateTitle(title: unknown): string {
-  if (typeof title !== 'string' || title.trim().length === 0) throw new Error('Title must not be empty')
-  const trimmed = title.trim()
-  if (trimmed.length > TITLE_MAX_LENGTH) throw new Error('Title too long')
-  return trimmed
-}
-
-function clampInt(value: number | undefined, min: number, max: number): number | undefined {
-  if (value == null) return undefined
-  const floored = Math.floor(value)
-  if (!Number.isFinite(floored)) return undefined
-  return Math.max(min, Math.min(max, floored))
-}
-
-function sanitizeFilenameBase(productName: string, fallback: string): string {
-  const source = (productName || fallback || 'analytics').replace(/[^a-zA-Z0-9_-]/g, '_')
-  return source.length > 0 ? source : 'analytics'
-}
-
-interface AnalyticsExportPreparation {
-  ok: true
-  exportData: Awaited<ReturnType<typeof buildAnalyticsExport>>
-}
-interface AnalyticsExportPreparationFail {
-  ok: false
-  error: string
-  /** When the failure happens before the export can be assembled
-   * (e.g. snapshot missing) we still surface the bits the dialog
-   * needs to render the validation card. Default 0 / 0 keeps the
-   * card showing red without blowing up. */
-  totalKeystrokes: number
-  rangeMs: number
-}
-
-/** Shared assembly path for both upload and preview. Reads the saved
- * filter snapshot, resolves the keymap snapshot main-side (snapshots
- * are local-only), folds the user's filter shape into the Hub's
- * `HubAnalyticsFilters` shape, and runs the builder. */
-async function prepareAnalyticsExport(
-  params: HubUploadAnalyticsPostParams | (HubPreviewAnalyticsPostParams & { title: string; thumbnailBase64: string }),
-): Promise<AnalyticsExportPreparation | AnalyticsExportPreparationFail> {
-  if (!params.uid || typeof params.uid !== 'string') {
-    return { ok: false, error: 'Invalid uid', totalKeystrokes: 0, rangeMs: 0 }
-  }
-  if (!params.entryId || typeof params.entryId !== 'string') {
-    return { ok: false, error: 'Invalid entryId', totalKeystrokes: 0, rangeMs: 0 }
-  }
-  const found = await readAnalyzeFilterEntry(params.uid, params.entryId)
-  if (!found) {
-    return { ok: false, error: 'Saved filter entry not found', totalKeystrokes: 0, rangeMs: 0 }
-  }
-
-  let payload: AnalyzeFilterSnapshotPayloadShape
-  try {
-    payload = JSON.parse(found.data) as AnalyzeFilterSnapshotPayloadShape
-  } catch {
-    return { ok: false, error: 'Saved filter payload is not valid JSON', totalKeystrokes: 0, rangeMs: 0 }
-  }
-  if (!payload || typeof payload !== 'object' || payload.version !== 1) {
-    return { ok: false, error: 'Unsupported saved filter version', totalKeystrokes: 0, rangeMs: 0 }
-  }
-  const range = payload.range
-  if (!range || typeof range.fromMs !== 'number' || typeof range.toMs !== 'number') {
-    return { ok: false, error: 'Saved filter has no range', totalKeystrokes: 0, rangeMs: 0 }
-  }
-  const rangeMs = Math.max(0, range.toMs - range.fromMs)
-
-  const deviceScope = resolveDeviceScopeFromPayload(payload.filters?.deviceScopes)
-  const appScopes = Array.isArray(payload.filters?.appScopes)
-    ? payload.filters.appScopes.filter((v): v is string => typeof v === 'string')
-    : []
-
-  // Snapshots are own-only — the typing-analytics service writes them
-  // against the local machine hash. The Analyze view itself reads the
-  // snapshot via `typingAnalyticsGetKeymapSnapshotForRange` which
-  // resolves the same hash internally.
-  const ownHash = await getMachineHash()
-  const snapshot = await getKeymapSnapshotForRange(
-    app.getPath('userData'), params.uid, ownHash, range.fromMs, range.toMs,
-  )
-  if (!snapshot) {
-    return { ok: false, error: 'No keymap snapshot recorded for this range', totalKeystrokes: 0, rangeMs }
-  }
-
-  const filters = projectFiltersForHub(payload, params.fingerOverrides)
-
-  const layoutInputs = params.layoutComparisonInputs
-  const layoutComparisonInputs: BuildAnalyticsExportInput['layoutComparisonInputs'] = layoutInputs
-    ? {
-        source: layoutInputs.source,
-        targets: layoutInputs.targets,
-        metrics: filterValidLayoutMetrics(layoutInputs.metrics),
-        kleKeys: layoutInputs.kleKeys as KleKey[],
-        layer: layoutInputs.layer,
-      }
-    : null
-
-  // Renderer-side category picker — only the listed sections get
-  // fetched. Unset / empty array ships everything (back-compat with
-  // the early build that did not surface the picker).
-  const categories = Array.isArray(params.categories) && params.categories.length > 0
-    ? new Set(params.categories.filter((c): c is HubAnalyticsCategoryId => typeof c === 'string'))
-    : undefined
-
-  const appDataApps = Array.isArray(params.appDataApps)
-    ? params.appDataApps.filter((v): v is string => typeof v === 'string')
-    : undefined
-
-  const exportData = await buildAnalyticsExport({
-    uid: params.uid,
-    productName: params.keyboard.productName,
-    vendorId: params.keyboard.vendorId,
-    productId: params.keyboard.productId,
-    snapshot: snapshot as TypingKeymapSnapshot,
-    range: { fromMs: range.fromMs, toMs: range.toMs },
-    deviceScope,
-    appScopes,
-    filters,
-    layoutComparisonInputs,
-    fingerOverrides: sanitizeFingerOverrides(params.fingerOverrides),
-    categories,
-    appDataApps,
-  })
-
-  return { ok: true, exportData }
-}
-
-/** Subset of the renderer-side AnalyzeFilterSnapshotPayload that the
- * main-side preparer needs to read. Re-stating the shape here keeps
- * the main module independent of the renderer-only hook file. */
-interface AnalyzeFilterSnapshotPayloadShape {
-  version: number
-  analysisTab?: string
-  range?: { fromMs?: number; toMs?: number }
-  filters?: {
-    deviceScopes?: unknown[]
-    appScopes?: unknown[]
-    heatmap?: Record<string, unknown>
-    wpm?: Record<string, unknown>
-    interval?: Record<string, unknown>
-    activity?: Record<string, unknown>
-    layer?: Record<string, unknown>
-    ergonomics?: Record<string, unknown>
-    bigrams?: Record<string, unknown>
-    layoutComparison?: Record<string, unknown>
-  }
-}
-
-function resolveDeviceScopeFromPayload(scopes: unknown): DeviceScope {
-  if (!Array.isArray(scopes) || scopes.length === 0) return 'own'
-  const first = scopes[0]
-  if (first === 'all' || first === 'own') return first
-  if (typeof first === 'object' && first !== null) {
-    const o = first as Record<string, unknown>
-    if (o.kind === 'hash' && typeof o.machineHash === 'string' && o.machineHash.length > 0) {
-      return { kind: 'hash', machineHash: o.machineHash }
-    }
-  }
-  return 'own'
-}
-
-const VALID_LAYOUT_METRICS: ReadonlySet<LayoutComparisonMetric> = new Set([
-  'fingerLoad', 'handBalance', 'rowDist', 'homeRow',
-])
-
-function filterValidLayoutMetrics(metrics: readonly string[] | undefined): LayoutComparisonMetric[] {
-  if (!Array.isArray(metrics)) return []
-  return metrics.filter((m): m is LayoutComparisonMetric => VALID_LAYOUT_METRICS.has(m as LayoutComparisonMetric))
-}
-
-function projectFiltersForHub(
-  payload: AnalyzeFilterSnapshotPayloadShape,
-  fingerOverrides: Record<string, string> | undefined,
-): HubAnalyticsFilters {
-  const f = payload.filters ?? {}
-  // Bigrams limits are fixed (10/10/20) per HUB-ANALYTICS-API.md §4.3
-  // — the desktop never sends user-tweaked counts so the Hub size /
-  // privacy surface stays predictable.
-  const pairThreshold = typeof f.bigrams?.pairIntervalThresholdMs === 'number'
-    ? f.bigrams.pairIntervalThresholdMs
-    : undefined
-  return {
-    // Always pin the Hub initial-tab hint to Summary so the post
-    // detail page lands on the at-a-glance view regardless of which
-    // tab was open when the saved condition was uploaded. Mirrors the
-    // local Load behaviour (handleLoadFilterSnapshot) which also
-    // forces Summary.
-    analysisTab: 'summary',
-    heatmap: f.heatmap,
-    wpm: f.wpm,
-    interval: f.interval,
-    activity: f.activity,
-    layer: f.layer,
-    ergonomics: f.ergonomics,
-    bigrams: { topLimit: 10, slowLimit: 10, fingerLimit: 20, pairIntervalThresholdMs: pairThreshold },
-    layoutComparison: f.layoutComparison,
-    fingerOverrides: fingerOverrides && Object.keys(fingerOverrides).length > 0 ? fingerOverrides : undefined,
-  }
-}
-
-function computeTotalPages(total: number, perPage: number): number {
-  const safeTotal = Number.isFinite(total) ? Math.max(0, total) : 0
-  const safePerPage = Number.isFinite(perPage) && perPage > 0 ? perPage : 1
-  return Math.max(1, Math.ceil(safeTotal / safePerPage))
-}
-
-// Cache Hub JWT to avoid redundant /api/auth/token round-trips.
-// Hub JWT is valid for 7 days; we cache for 24 hours.
-// withTokenRetry() handles mid-cache expiry via automatic 401 retry.
-// The /api/auth/token endpoint has a 10 req/min rate limit.
-const HUB_JWT_TTL_MS = 24 * 60 * 60 * 1000
-let cachedHubJwt: { token: string; expiresAt: number } | null = null
-let inflightHubAuth: Promise<string> | null = null
-let cacheGeneration = 0
-let pendingAuthDisplayName: string | null = null
-
-async function getHubToken(): Promise<string> {
-  if (cachedHubJwt && Date.now() < cachedHubJwt.expiresAt) {
-    return cachedHubJwt.token
-  }
-  // Deduplicate concurrent requests
-  if (inflightHubAuth) return inflightHubAuth
-  const gen = cacheGeneration
-  inflightHubAuth = (async () => {
-    try {
-      const idToken = await getIdToken()
-      if (!idToken) throw new Error(AUTH_ERROR)
-      let auth: HubAuthResult
-      try {
-        auth = await authenticateWithHub(idToken, pendingAuthDisplayName ?? undefined)
-      } catch (err) {
-        if (err instanceof Hub409Error) throw new Error(HUB_ERROR_DISPLAY_NAME_CONFLICT)
-        rethrowAsHubSentinel(err)
-        throw err
-      }
-      // Only cache if not invalidated (e.g. by sign-out) during the request
-      if (gen === cacheGeneration) {
-        cachedHubJwt = { token: auth.token, expiresAt: Date.now() + HUB_JWT_TTL_MS }
-      }
-      return auth.token
-    } finally {
-      inflightHubAuth = null
-    }
-  })()
-  return inflightHubAuth
-}
-
-export function clearHubTokenCache(): void {
-  cachedHubJwt = null
-  inflightHubAuth = null
-  cacheGeneration++
-  pendingAuthDisplayName = null
-}
-
-function invalidateCachedHubJwt(): void {
-  cachedHubJwt = null
-}
-
-function extractError(err: unknown, fallback: string): string {
-  return err instanceof Error ? err.message : fallback
-}
-
-/** Rebuilds a failed lookup result under a different `T` — `getRecord`
- * resolves `KeyLabelStoreResult<KeyLabelRecord>`, but the handler needs to
- * return `KeyLabelStoreResult<KeyLabelMeta>`, and the two `T`s don't share
- * a `data` shape, so `record as KeyLabelStoreResult<KeyLabelMeta>` would be
- * an unsound cast. Only the failure fields (`success`/`errorCode`/`error`)
- * carry over. */
-function toKeyLabelLookupFailure(record: {
-  success: boolean
-  errorCode?: KeyLabelStoreResult<unknown>['errorCode']
-  error?: string
-}): KeyLabelStoreResult<KeyLabelMeta> {
-  return { success: record.success, errorCode: record.errorCode, error: record.error }
-}
-
-function toPrivateResult(res: { id: string; url: string; expires_at: string | null }): HubPrivateUploadResult {
-  return { success: true, id: res.id, url: res.url, expiresAt: res.expires_at }
-}
-
-function rethrowAsHubSentinel(err: unknown): void {
-  if (err instanceof Hub403Error) throw new Error(HUB_ERROR_ACCOUNT_DEACTIVATED)
-  if (err instanceof Hub429Error) throw new Error(HUB_ERROR_RATE_LIMITED)
-}
-
-async function withTokenRetry<T>(operation: (jwt: string) => Promise<T>): Promise<T> {
-  const jwt = await getHubToken()
-  try {
-    return await operation(jwt)
-  } catch (err) {
-    if (err instanceof Hub401Error) {
-      invalidateCachedHubJwt()
-      const freshJwt = await getHubToken()
-      try {
-        return await operation(freshJwt)
-      } catch (retryErr) {
-        rethrowAsHubSentinel(retryErr)
-        throw retryErr
-      }
-    }
-    rethrowAsHubSentinel(err)
-    throw err
-  }
-}
-
-const MB = 1024 * 1024
-const FILE_SIZE_LIMITS: Record<string, { max: number; label: string }> = {
-  thumbnail: { max: 2 * MB, label: 'thumbnail' },
-  vil: { max: 10 * MB, label: 'vil' },
-  pipette: { max: 10 * MB, label: 'pipette' },
-  c: { max: 10 * MB, label: 'keymap C' },
-  pdf: { max: 10 * MB, label: 'PDF' },
-}
-
-function validateFileSize(files: HubUploadFiles): void {
-  for (const [key, limit] of Object.entries(FILE_SIZE_LIMITS)) {
-    const file = files[key as keyof HubUploadFiles]
-    if (file.data.byteLength > limit.max) {
-      throw new Error(`File too large: ${limit.label} exceeds ${limit.max / MB} MB limit`)
-    }
-  }
-}
-
-function buildFiles(params: HubUploadPostParams): HubUploadFiles {
-  const baseName = params.keyboardName.replace(/[^a-zA-Z0-9_-]/g, '_')
-  const files: HubUploadFiles = {
-    vil: { name: `${baseName}.vil`, data: Buffer.from(params.vilJson, 'utf-8') },
-    pipette: { name: `${baseName}.pipette`, data: Buffer.from(params.pipetteJson, 'utf-8') },
-    c: { name: `${baseName}.c`, data: Buffer.from(params.keymapC, 'utf-8') },
-    pdf: { name: `${baseName}.pdf`, data: Buffer.from(params.pdfBase64, 'base64') },
-    thumbnail: { name: `${baseName}.jpg`, data: Buffer.from(params.thumbnailBase64, 'base64') },
-  }
-  validateFileSize(files)
-  return files
-}
-
-// Deliberately stricter than the shared `isSafePathSegment` (ASCII
-// allowlist vs separator denylist) because these filenames cross the
-// Hub upload boundary; not consolidated by design.
-function isSafeExportFilename(segment: string): boolean {
-  return /^[a-zA-Z0-9_.-]+$/.test(segment) && segment !== '.' && segment !== '..'
-}
-
-async function buildFavoriteExportJson(
-  type: FavoriteType,
-  entryId: string,
-  vialProtocol: number,
-): Promise<string> {
-  const favDir = join(app.getPath('userData'), 'sync', 'favorites', type)
-  const indexPath = join(favDir, 'index.json')
-  const indexRaw = await readFile(indexPath, 'utf-8')
-  const index = JSON.parse(indexRaw) as FavoriteIndex
-  const entry = index.entries.find((e) => e.id === entryId && !e.deletedAt)
-  if (!entry) throw new Error('Entry not found')
-
-  if (!isSafeExportFilename(entry.filename)) throw new Error('Invalid filename')
-  const filePath = join(favDir, entry.filename)
-  const fileRaw = await readFile(filePath, 'utf-8')
-  const parsed = JSON.parse(fileRaw) as { type: string; data: unknown }
-  if (parsed.data == null) throw new Error('Entry data is empty')
-  if (parsed.type !== type) throw new Error('Entry type mismatch')
-
-  const exportKey = FAV_TYPE_TO_EXPORT_KEY[type]
-  const serializedData = withSerializeProtocol(vialProtocol, () => serializeFavData(type, parsed.data, serializeKeycode))
-
-  const exportFile = buildFavExportFile(vialProtocol, {
-    [exportKey]: [{
-      label: entry.label,
-      savedAt: entry.savedAt,
-      data: serializedData,
-    }],
-  })
-
-  return JSON.stringify(exportFile)
-}
-
-/**
- * Fetch a key-label download body together with the uploader name and
- * Hub-side `updated_at` from the detail endpoint. The detail call is
- * best-effort: if it fails (network blip, 404 on a deleted post, etc.)
- * we keep the caller-supplied uploader fallback so the Author column
- * does not lose its cached value, and `hubUpdatedAt` simply stays
- * undefined for that round.
- */
-async function fetchHubKeyLabelPayload(
-  hubPostId: string,
-  fallbackUploader?: string,
-  fallbackHubUpdatedAt?: string,
-): Promise<{
-  body: {
-    name: string
-    map: Record<string, string>
-    composite_labels: Record<string, string> | null
-    keymap_applicable?: boolean
-  }
-  uploaderName: string | undefined
-  hubUpdatedAt: string | undefined
-}> {
-  const body = await downloadKeyLabel(hubPostId)
-  let uploaderName: string | undefined = fallbackUploader
-  let hubUpdatedAt: string | undefined = fallbackHubUpdatedAt
-  try {
-    const detail = await fetchKeyLabelDetail(hubPostId)
-    uploaderName = detail.uploader_name ?? fallbackUploader
-    hubUpdatedAt = detail.updated_at ?? fallbackHubUpdatedAt
-  } catch {
-    // best-effort; keep the fallback values
-  }
-  return { body, uploaderName, hubUpdatedAt }
-}
-
-/**
- * `HubKeyLabelInput` shared by upload/update — the local record's name,
- * map and compositeLabels, plus `keymapApplicable` (always sent, even
- * `false`, so a re-upload can clear a previously-true flag; see the
- * `buildBody` comment in hub-key-labels.ts).
- */
-function buildHubKeyLabelInput(record: KeyLabelRecord): HubKeyLabelInput {
-  return {
-    name: record.meta.name,
-    map: record.data.map,
-    ...(record.data.compositeLabels ? { compositeLabels: record.data.compositeLabels } : {}),
-    keymapApplicable: record.data.keymapApplicable ?? false,
-  }
-}
-
-/**
- * `saveRecord()` fields shared by download/sync — the Hub body's map,
- * compositeLabels and keymapApplicable, plus the cached uploader /
- * hub-updated metadata. Callers still supply `id`/`name`/`hubPostId`
- * since those differ (download seeds from the Hub body, sync keeps the
- * local id/name and only refreshes the payload).
- */
-function hubBodyToSaveRecordFields(
-  body: { map: Record<string, string>; composite_labels: Record<string, string> | null; keymap_applicable?: boolean },
-  uploaderName: string | undefined,
-  hubUpdatedAt: string | undefined,
-): Pick<SaveRecordInput, 'map' | 'uploaderName' | 'compositeLabels' | 'keymapApplicable' | 'hubUpdatedAt'> {
-  const composite = body.composite_labels ?? undefined
-  return {
-    map: body.map,
-    ...(uploaderName ? { uploaderName } : {}),
-    ...(composite ? { compositeLabels: composite } : {}),
-    ...(body.keymap_applicable ? { keymapApplicable: true } : {}),
-    ...(hubUpdatedAt ? { hubUpdatedAt } : {}),
-  }
-}
-
-/**
- * i18n/theme upload responses only carry `{ id, title }` (`HubPostResponse`)
- * — unlike Key Labels' upload/update responses, which return the full
- * `HubKeyLabelItem` including `uploader_name`/`updated_at` directly.
- * Neither `/api/i18n-packs` nor `/api/theme-packs` expose a single-item
- * detail GET to fill that gap, so this re-uses the existing exact-name
- * list filter (`?name=`) to find the just-created/updated item and read
- * its `uploaderName`/`updatedAt` off the list response instead. Matches
- * `fetchHubKeyLabelPayload`'s best-effort contract: on any failure (or if
- * the id is not found in that name's results) both fields come back
- * `undefined` and the caller's `setHubPostId` call simply leaves the
- * previously-cached values alone.
- *
- * Upload is the only caller: a freshly created post needs `uploaderName`
- * (the Author column has nothing else to key off yet). Update deliberately
- * uses the narrower `enrichHubUpdatedAt` below instead of this — Update
- * never changes `uploaderName`, so paying for the name-filtered list
- * lookup (paginated, larger body, and — in the pathological case of two
- * packs sharing a name — order-dependent) would buy nothing over the
- * id-keyed timestamps endpoint. Both helpers cost the same +1 Hub
- * request per action either way; the split is about which endpoint
- * shape fits what the caller actually needs, not request count.
- */
-async function enrichHubPackMeta(
-  listByName: (params: { name: string }) => Promise<{ items: { id: string; uploaderName?: string | null; updatedAt?: string }[] }>,
-  name: string,
-  postId: string,
-): Promise<{ uploaderName?: string; hubUpdatedAt?: string }> {
-  try {
-    const list = await listByName({ name })
-    const item = list.items.find((i) => i.id === postId)
-    return {
-      uploaderName: item?.uploaderName ?? undefined,
-      hubUpdatedAt: item?.updatedAt ?? undefined,
-    }
-  } catch {
-    return {}
-  }
-}
-
-/**
- * Update-only counterpart to `enrichHubPackMeta`: since Update never
- * touches `uploaderName` (see the i18n/theme Update handlers below), it
- * only ever needs `hubUpdatedAt`. The id-keyed `/timestamps` endpoint is
- * the better fit for that — no pagination, no name-collision ambiguity,
- * and a much smaller response body than the full list lookup. Same
- * best-effort contract as `enrichHubPackMeta`: any failure returns `{}`
- * and the caller's `setHubPostId` leaves the cached `hubUpdatedAt` alone.
- */
-async function enrichHubUpdatedAt(
-  fetchTimestamps: (ids: string[]) => Promise<{ items: { id: string; updated_at: string }[] }>,
-  postId: string,
-): Promise<{ hubUpdatedAt?: string }> {
-  try {
-    const res = await fetchTimestamps([postId])
-    return { hubUpdatedAt: res.items.find((i) => i.id === postId)?.updated_at }
-  } catch {
-    return {}
-  }
-}
+import { registerHubPostHandlers } from './hub-ipc-posts'
+import { registerHubFavoriteHandlers } from './hub-ipc-favorite'
+import { registerHubAnalyticsHandlers } from './hub-ipc-analytics'
+import { registerHubPrivateHandlers } from './hub-ipc-private'
+import { registerHubPackHandlers } from './hub-ipc-packs'
+import { registerHubKeyLabelHandlers } from './hub-ipc-key-labels'
 
 export function setupHubIpc(): void {
-  secureHandle(
-    IpcChannels.HUB_UPLOAD_POST,
-    async (_event, params: HubUploadPostParams): Promise<HubUploadResult> => {
-      try {
-        const title = validateTitle(params.title)
-        const files = buildFiles(params)
-        const result = await withTokenRetry((jwt) =>
-          uploadPostToHub(jwt, title, params.keyboardName, files),
-        )
-        return { success: true, postId: result.id }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Upload failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_UPDATE_POST,
-    async (_event, params: HubUpdatePostParams): Promise<HubUploadResult> => {
-      try {
-        validatePostId(params.postId)
-        const title = validateTitle(params.title)
-        const files = buildFiles(params)
-        const result = await withTokenRetry((jwt) =>
-          updatePostOnHub(jwt, params.postId, title, params.keyboardName, files),
-        )
-        return { success: true, postId: result.id }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Update failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_PATCH_POST,
-    async (_event, params: HubPatchPostParams): Promise<HubDeleteResult> => {
-      try {
-        validatePostId(params.postId)
-        const title = validateTitle(params.title)
-        await withTokenRetry((jwt) =>
-          patchPostOnHub(jwt, params.postId, { title }),
-        )
-        return { success: true }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Patch failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_DELETE_POST,
-    async (_event, postId: string): Promise<HubDeleteResult> => {
-      try {
-        validatePostId(postId)
-        await withTokenRetry((jwt) => deletePostFromHub(jwt, postId))
-        return { success: true }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Delete failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_FETCH_MY_POSTS,
-    async (_event, params?: HubFetchMyPostsParams): Promise<HubFetchMyPostsResult> => {
-      try {
-        const page = clampInt(params?.page, 1, Number.MAX_SAFE_INTEGER)
-        const perPage = clampInt(params?.per_page, 1, 100)
-        const result = await withTokenRetry((jwt) =>
-          fetchMyPosts(jwt, { page, per_page: perPage }),
-        )
-        const totalPages = computeTotalPages(result.total, result.per_page)
-        return {
-          success: true,
-          posts: result.items,
-          pagination: {
-            total: result.total,
-            page: result.page,
-            per_page: result.per_page,
-            total_pages: totalPages,
-          },
-        }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Fetch my posts failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_FETCH_AUTH_ME,
-    async (): Promise<HubUserResult> => {
-      try {
-        const user = await withTokenRetry((jwt) => fetchAuthMe(jwt))
-        return { success: true, user }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Fetch auth failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_PATCH_AUTH_ME,
-    async (_event, displayName: unknown): Promise<HubUserResult> => {
-      try {
-        const validated = validateDisplayName(displayName)
-        const user = await withTokenRetry((jwt) => patchAuthMe(jwt, validated))
-        return { success: true, user }
-      } catch (err) {
-        if (err instanceof Hub409Error) {
-          return { success: false, error: HUB_ERROR_DISPLAY_NAME_CONFLICT }
-        }
-        return { success: false, error: extractError(err, 'Patch auth failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_FETCH_MY_KEYBOARD_POSTS,
-    async (_event, keyboardName: unknown): Promise<HubFetchMyKeyboardPostsResult> => {
-      try {
-        const validated = validateKeyboardName(keyboardName)
-        const posts = await withTokenRetry((jwt) =>
-          fetchMyPostsByKeyboard(jwt, validated),
-        )
-        return { success: true, posts }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Fetch keyboard posts failed') }
-      }
-    },
-  )
-
-  secureHandle(IpcChannels.HUB_GET_ORIGIN, (): string => getHubOrigin())
-
-  secureHandle(
-    IpcChannels.HUB_SET_AUTH_DISPLAY_NAME,
-    (_event, displayName: string | null): void => {
-      pendingAuthDisplayName = typeof displayName === 'string' ? displayName : null
-      // Invalidate cached JWT so the next getHubToken() re-authenticates
-      // with the new display name instead of returning a stale cached/inflight result.
-      cachedHubJwt = null
-      inflightHubAuth = null
-    },
-  )
-
-  // --- Favorite (feature) post handlers ---
-
-  async function prepareFavoritePost(
-    params: HubUploadFavoritePostParams,
-  ): Promise<{ title: string; postType: string; jsonFile: { name: string; data: Buffer } }> {
-    if (!isValidFavoriteType(params.type)) throw new Error('Invalid favorite type')
-    // Hub-specific: stricter than the general `isValidVialProtocol` (see
-    // favorite-data.ts) because a raw negative sentinel (e.g. -1, the
-    // disconnected-state placeholder) must fail fast locally with a clear
-    // error instead of reaching the Hub server as a confusing 400.
-    if (!isValidHubVialProtocol(params.vialProtocol)) throw new Error('Invalid vialProtocol')
-    const title = validateTitle(params.title)
-    const postType = FAV_TYPE_TO_EXPORT_KEY[params.type]
-    const jsonStr = await buildFavoriteExportJson(params.type, params.entryId, params.vialProtocol)
-    return { title, postType, jsonFile: { name: `${postType}.json`, data: Buffer.from(jsonStr, 'utf-8') } }
-  }
-
-  secureHandle(
-    IpcChannels.HUB_UPLOAD_FAVORITE_POST,
-    async (_event, params: HubUploadFavoritePostParams): Promise<HubUploadResult> => {
-      try {
-        const { title, postType, jsonFile } = await prepareFavoritePost(params)
-        const result = await withTokenRetry((jwt) =>
-          uploadFeaturePostToHub(jwt, title, postType, jsonFile),
-        )
-        return { success: true, postId: result.id }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Upload failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_UPDATE_FAVORITE_POST,
-    async (_event, params: HubUpdateFavoritePostParams): Promise<HubUploadResult> => {
-      try {
-        validatePostId(params.postId)
-        const { title, postType, jsonFile } = await prepareFavoritePost(params)
-        const result = await withTokenRetry((jwt) =>
-          updateFeaturePostOnHub(jwt, params.postId, title, postType, jsonFile),
-        )
-        return { success: true, postId: result.id }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Update failed') }
-      }
-    },
-  )
-
-  // --- Analytics post handlers ---
-  //
-  // Pattern mirrors the favorite-post upload: validate inputs → assemble
-  // payload → withTokenRetry → save the postId on success. Distinct
-  // because the analytics build step is heavier (fetches across the
-  // typing-analytics DB) and ships a thumbnail alongside the JSON.
-
-  secureHandle(
-    IpcChannels.HUB_UPLOAD_ANALYTICS_POST,
-    async (_event, params: HubUploadAnalyticsPostParams): Promise<HubUploadResult> => {
-      try {
-        const built = await prepareAnalyticsExport(params)
-        if (!built.ok) return { success: false, error: built.error }
-        const title = validateTitle(params.title)
-        const baseName = sanitizeFilenameBase(params.keyboard.productName, params.uid)
-        const jsonBuffer = Buffer.from(JSON.stringify(built.exportData), 'utf-8')
-        const thumbnailBuffer = Buffer.from(params.thumbnailBase64, 'base64')
-        const result = await withTokenRetry((jwt) =>
-          uploadAnalyticsPostToHub(
-            jwt,
-            title,
-            { name: `${baseName}.json`, data: jsonBuffer },
-            { name: `${baseName}.jpg`, data: thumbnailBuffer },
-          ),
-        )
-        // Save the postId synchronously after upload so the panel can
-        // immediately show the "↻ Hub" / 🔗 affordances without a
-        // round-trip; failures here don't undo the upload (the entry
-        // would just appear unsynced and the next click would attempt
-        // a fresh upload).
-        await setAnalyzeFilterHubPostId(params.uid, params.entryId, result.id)
-        return { success: true, postId: result.id }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Analytics upload failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_UPDATE_ANALYTICS_POST,
-    async (_event, params: HubUpdateAnalyticsPostParams): Promise<HubUploadResult> => {
-      try {
-        validatePostId(params.postId)
-        const built = await prepareAnalyticsExport(params)
-        if (!built.ok) return { success: false, error: built.error }
-        const title = validateTitle(params.title)
-        const baseName = sanitizeFilenameBase(params.keyboard.productName, params.uid)
-        const jsonBuffer = Buffer.from(JSON.stringify(built.exportData), 'utf-8')
-        const thumbnailBuffer = Buffer.from(params.thumbnailBase64, 'base64')
-        const result = await withTokenRetry((jwt) =>
-          updateAnalyticsPostOnHub(
-            jwt,
-            params.postId,
-            title,
-            { name: `${baseName}.json`, data: jsonBuffer },
-            { name: `${baseName}.jpg`, data: thumbnailBuffer },
-          ),
-        )
-        // Re-stamp the postId in case the user manipulated the saved
-        // entry's metadata in another window between preview and
-        // upload — keeps the local index in sync with the Hub canon.
-        await setAnalyzeFilterHubPostId(params.uid, params.entryId, result.id)
-        return { success: true, postId: result.id }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Analytics update failed') }
-      }
-    },
-  )
-
-  // Preview path used by the upload dialog. Builds the payload with
-  // the same builder the upload uses but reports size + validation
-  // without crossing the network. The thumbnail is captured later
-  // (only when the user confirms), so it's intentionally absent from
-  // the preview params.
-  secureHandle(
-    IpcChannels.HUB_PREVIEW_ANALYTICS_POST,
-    async (_event, params: HubPreviewAnalyticsPostParams): Promise<{ success: boolean; preview?: HubAnalyticsPreview; error?: string }> => {
-      try {
-        const built = await prepareAnalyticsExport({
-          ...params,
-          // The preview path doesn't ship a thumbnail — pass empty
-          // strings to satisfy the shared param shape without
-          // triggering the buffer encode for nothing.
-          title: 'preview',
-          thumbnailBase64: '',
-        })
-        if (!built.ok) {
-          return {
-            success: true,
-            preview: {
-              totalKeystrokes: built.totalKeystrokes,
-              rangeMs: built.rangeMs,
-              estimatedBytes: 0,
-              validation: { ok: false, reason: built.error },
-            },
-          }
-        }
-        const validation = validateAnalyticsExport(built.exportData)
-        const estimatedBytes = estimateAnalyticsExportSizeBytes(built.exportData)
-        return {
-          success: true,
-          preview: {
-            totalKeystrokes: built.exportData.snapshot.totalKeystrokes,
-            rangeMs: built.exportData.snapshot.range.toMs - built.exportData.snapshot.range.fromMs,
-            estimatedBytes,
-            validation,
-          },
-        }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Analytics preview failed') }
-      }
-    },
-  )
-
-  // --- i18n language pack handlers ---
-  //
-  // Targets the standalone /api/i18n-packs endpoints. The pack body
-  // (HubI18nPackBody) carries name + version, so the multipart upload
-  // does not need a separate title field — the Hub derives identity
-  // from `pack.name` + `pack.version`. Listing / download stays
-  // anonymous; CRUD requires the JWT.
-
-  // --- Private (unlisted) upload handlers ---
-  //
-  // Each uploads to `/api/private/*` and returns the relative share URL
-  // + expiry. Persisting the link onto the local entry is the renderer's
-  // job (via the per-store `set-hub-private` IPC), mirroring how the
-  // public keymap/favorite uploads return a postId for the renderer to
-  // persist. The single delete handler is shared by Remove and the
-  // visibility-switch path; a 404 (already expired) counts as success.
-
-  secureHandle(
-    IpcChannels.HUB_UPLOAD_PRIVATE_POST,
-    async (_event, params: HubPrivateUploadPostParams): Promise<HubPrivateUploadResult> => {
-      try {
-        const title = validateTitle(params.title)
-        const files = buildFiles(params)
-        const res = await withTokenRetry((jwt) =>
-          uploadPrivatePostToHub(jwt, title, params.keyboardName, files, params.expiresInDays),
-        )
-        return toPrivateResult(res)
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Upload failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_UPLOAD_PRIVATE_FAVORITE_POST,
-    async (_event, params: HubPrivateUploadFavoritePostParams): Promise<HubPrivateUploadResult> => {
-      try {
-        const { title, postType, jsonFile } = await prepareFavoritePost(params)
-        const res = await withTokenRetry((jwt) =>
-          uploadPrivateFeaturePostToHub(jwt, title, postType, jsonFile, params.expiresInDays),
-        )
-        return toPrivateResult(res)
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Upload failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_UPLOAD_PRIVATE_ANALYTICS_POST,
-    async (_event, params: HubPrivateUploadAnalyticsPostParams): Promise<HubPrivateUploadResult> => {
-      try {
-        const built = await prepareAnalyticsExport(params)
-        if (!built.ok) return { success: false, error: built.error }
-        const title = validateTitle(params.title)
-        const baseName = sanitizeFilenameBase(params.keyboard.productName, params.uid)
-        const jsonBuffer = Buffer.from(JSON.stringify(built.exportData), 'utf-8')
-        const thumbnailBuffer = Buffer.from(params.thumbnailBase64, 'base64')
-        const res = await withTokenRetry((jwt) =>
-          uploadPrivateAnalyticsPostToHub(
-            jwt,
-            title,
-            { name: `${baseName}.json`, data: jsonBuffer },
-            { name: `${baseName}.jpg`, data: thumbnailBuffer },
-            params.expiresInDays,
-          ),
-        )
-        return toPrivateResult(res)
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Analytics upload failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_DELETE_PRIVATE_POST,
-    async (_event, kind: HubPrivateKind, id: string): Promise<HubDeleteResult> => {
-      try {
-        if (typeof id !== 'string' || !id) return { success: false, error: 'Invalid id' }
-        await withTokenRetry((jwt) => deletePrivatePostFromHub(jwt, kind, id))
-        return { success: true }
-      } catch (err) {
-        // Already gone (expired / removed elsewhere) — treat as success so
-        // the local link can be cleared / a visibility switch can proceed.
-        if (err instanceof Hub404Error) return { success: true }
-        return { success: false, error: extractError(err, 'Delete failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_UPLOAD_I18N_POST,
-    async (_event, params: HubUploadI18nPostParams): Promise<HubUploadResult> => {
-      try {
-        if (!params || typeof params !== 'object') return { success: false, error: 'Invalid params' }
-        const result = await withTokenRetry((jwt) => uploadI18nPostToHub(jwt, params.pack))
-        // Carry the response's uploader/updated_at into the local meta
-        // (via a name-matched list lookup — see enrichHubPackMeta) so
-        // the Author/Updated columns and the isMine gate populate
-        // immediately without waiting for a sync.
-        const enriched = await enrichHubPackMeta(fetchI18nPostList, params.pack.name, result.id)
-        await setI18nPackHubPostId(params.entryId, result.id, enriched.uploaderName, enriched.hubUpdatedAt)
-        return { success: true, postId: result.id }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'i18n upload failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_UPDATE_I18N_POST,
-    async (_event, params: HubUpdateI18nPostParams): Promise<HubUploadResult> => {
-      try {
-        if (!params || typeof params !== 'object') return { success: false, error: 'Invalid params' }
-        validatePostId(params.postId)
-        const result = await withTokenRetry((jwt) => updateI18nPostOnHub(jwt, params.postId, params.pack))
-        // Refresh hubUpdatedAt only — mirrors Key Labels' Update, which
-        // deliberately leaves uploaderName alone (the owner performing
-        // an update is assumed unchanged). Uses the id-keyed timestamps
-        // endpoint rather than enrichHubPackMeta's name-filtered list
-        // lookup, since uploaderName isn't needed here.
-        const enriched = await enrichHubUpdatedAt(fetchI18nPackTimestamps, result.id)
-        await setI18nPackHubPostId(params.entryId, result.id, undefined, enriched.hubUpdatedAt)
-        return { success: true, postId: result.id }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'i18n update failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_DELETE_I18N_POST,
-    async (_event, postId: unknown, localPackId: unknown): Promise<HubDeleteResult> => {
-      try {
-        if (typeof postId !== 'string') return { success: false, error: 'Invalid post ID' }
-        validatePostId(postId)
-        await withTokenRetry((jwt) => deleteI18nPostFromHub(jwt, postId))
-        if (typeof localPackId === 'string' && localPackId) {
-          await setI18nPackHubPostId(localPackId, null)
-        }
-        return { success: true }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'i18n delete failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_LIST_I18N_POSTS,
-    async (_event, params: unknown): Promise<{ success: boolean; data?: HubI18nListResponse; error?: string }> => {
-      try {
-        const query: HubI18nListParams = {}
-        if (params && typeof params === 'object') {
-          const obj = params as Record<string, unknown>
-          if (typeof obj.q === 'string') query.q = obj.q
-          if (typeof obj.name === 'string') query.name = obj.name
-          if (typeof obj.page === 'number') query.page = obj.page
-          if (typeof obj.perPage === 'number') query.perPage = obj.perPage
-        }
-        const data = await fetchI18nPostList(query)
-        return { success: true, data }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'i18n list failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_DOWNLOAD_I18N_POST,
-    async (_event, postId: unknown): Promise<{ success: boolean; data?: HubI18nExportV1; error?: string }> => {
-      try {
-        if (typeof postId !== 'string') return { success: false, error: 'Invalid post ID' }
-        validatePostId(postId)
-        const exportData = await downloadI18nPostBody(postId)
-        const validation = validateI18nExport(exportData)
-        if (!validation.ok) {
-          return { success: false, error: `Hub returned an invalid pack: ${validation.reason ?? 'unknown reason'}` }
-        }
-        return { success: true, data: exportData }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'i18n download failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_I18N_PACK_TIMESTAMPS,
-    async (_event, ids: unknown): Promise<{ success: boolean; data?: HubI18nPackTimestampsResponse; error?: string }> => {
-      if (!Array.isArray(ids) || !ids.every((id) => typeof id === 'string')) {
-        return { success: false, error: 'ids must be an array of strings' }
-      }
-      const unique = Array.from(new Set(ids as string[]))
-      if (unique.length === 0) return { success: true, data: { items: [] } }
-      try {
-        const chunks: string[][] = []
-        for (let i = 0; i < unique.length; i += HUB_I18N_PACK_TIMESTAMPS_BATCH_LIMIT) {
-          chunks.push(unique.slice(i, i + HUB_I18N_PACK_TIMESTAMPS_BATCH_LIMIT))
-        }
-        const responses = await Promise.all(chunks.map((chunk) => fetchI18nPackTimestamps(chunk)))
-        const byId = new Map<string, HubI18nPackTimestamp>()
-        for (const r of responses) {
-          for (const item of r.items) byId.set(item.id, item)
-        }
-        const items: HubI18nPackTimestamp[] = []
-        for (const id of unique) {
-          const found = byId.get(id)
-          if (found) items.push(found)
-        }
-        return { success: true, data: { items } }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Hub i18n timestamps failed') }
-      }
-    },
-  )
-
-  // --- Hub theme pack handlers ---
-
-  secureHandle(
-    IpcChannels.HUB_UPLOAD_THEME_POST,
-    async (_event, params: HubUploadThemePostParams): Promise<HubUploadResult> => {
-      try {
-        if (!params || typeof params !== 'object') return { success: false, error: 'Invalid params' }
-        const result = await withTokenRetry((jwt) => uploadThemePostToHub(jwt, params.pack))
-        const enriched = await enrichHubPackMeta(fetchThemePostList, params.pack.name, result.id)
-        await setThemePackHubPostId(params.entryId, result.id, enriched.uploaderName, enriched.hubUpdatedAt)
-        return { success: true, postId: result.id }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'theme upload failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_UPDATE_THEME_POST,
-    async (_event, params: HubUpdateThemePostParams): Promise<HubUploadResult> => {
-      try {
-        if (!params || typeof params !== 'object') return { success: false, error: 'Invalid params' }
-        validatePostId(params.postId)
-        const result = await withTokenRetry((jwt) => updateThemePostOnHub(jwt, params.postId, params.pack))
-        // hubUpdatedAt only, via the id-keyed timestamps endpoint —
-        // see the i18n Update handler's comment.
-        const enriched = await enrichHubUpdatedAt(fetchThemePackTimestamps, result.id)
-        await setThemePackHubPostId(params.entryId, result.id, undefined, enriched.hubUpdatedAt)
-        return { success: true, postId: result.id }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'theme update failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_DELETE_THEME_POST,
-    async (_event, postId: unknown, localPackId: unknown): Promise<HubDeleteResult> => {
-      try {
-        if (typeof postId !== 'string') return { success: false, error: 'Invalid post ID' }
-        validatePostId(postId)
-        await withTokenRetry((jwt) => deleteThemePostFromHub(jwt, postId))
-        if (typeof localPackId === 'string' && localPackId) {
-          await setThemePackHubPostId(localPackId, null)
-        }
-        return { success: true }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'theme delete failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_LIST_THEME_POSTS,
-    async (_event, params: unknown): Promise<{ success: boolean; data?: HubThemeListResponse; error?: string }> => {
-      try {
-        const query: HubThemeListParams = {}
-        if (params && typeof params === 'object') {
-          const obj = params as Record<string, unknown>
-          if (typeof obj.q === 'string') query.q = obj.q
-          if (typeof obj.name === 'string') query.name = obj.name
-          if (typeof obj.page === 'number') query.page = obj.page
-          if (typeof obj.perPage === 'number') query.perPage = obj.perPage
-        }
-        const data = await fetchThemePostList(query)
-        return { success: true, data }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'theme list failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_DOWNLOAD_THEME_POST,
-    async (_event, postId: unknown): Promise<{ success: boolean; data?: HubThemePackBody; error?: string }> => {
-      try {
-        if (typeof postId !== 'string') return { success: false, error: 'Invalid post ID' }
-        validatePostId(postId)
-        const packBody = await downloadThemePostBody(postId)
-        const validation = validateThemePack(packBody)
-        if (!validation.ok) {
-          return { success: false, error: `Hub returned an invalid theme pack: ${validation.errors[0] ?? 'unknown reason'}` }
-        }
-        return { success: true, data: packBody }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'theme download failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.HUB_THEME_PACK_TIMESTAMPS,
-    async (_event, ids: unknown): Promise<{ success: boolean; data?: HubThemePackTimestampsResponse; error?: string }> => {
-      if (!Array.isArray(ids) || !ids.every((id) => typeof id === 'string')) {
-        return { success: false, error: 'ids must be an array of strings' }
-      }
-      const unique = Array.from(new Set(ids as string[]))
-      if (unique.length === 0) return { success: true, data: { items: [] } }
-      try {
-        const chunks: string[][] = []
-        for (let i = 0; i < unique.length; i += HUB_THEME_PACK_TIMESTAMPS_BATCH_LIMIT) {
-          chunks.push(unique.slice(i, i + HUB_THEME_PACK_TIMESTAMPS_BATCH_LIMIT))
-        }
-        const responses = await Promise.all(chunks.map((chunk) => fetchThemePackTimestamps(chunk)))
-        const byId = new Map<string, HubThemePackTimestamp>()
-        for (const r of responses) {
-          for (const item of r.items) byId.set(item.id, item)
-        }
-        const items: HubThemePackTimestamp[] = []
-        for (const id of unique) {
-          const found = byId.get(id)
-          if (found) items.push(found)
-        }
-        return { success: true, data: { items } }
-      } catch (err) {
-        return { success: false, error: extractError(err, 'Hub theme timestamps failed') }
-      }
-    },
-  )
-
-  // --- Key Label Hub handlers ---
-
-  secureHandle(
-    IpcChannels.KEY_LABEL_HUB_LIST,
-    async (
-      _event,
-      params: HubKeyLabelListParams | undefined,
-    ): Promise<KeyLabelStoreResult<HubKeyLabelListResponse>> => {
-      try {
-        const page = clampInt(params?.page, 1, Number.MAX_SAFE_INTEGER) ?? 1
-        const perPage = clampInt(params?.perPage, 1, 100) ?? 20
-        const data = await fetchKeyLabelList({ q: params?.q, page, perPage })
-        return { success: true, data }
-      } catch (err) {
-        return { success: false, errorCode: 'IO_ERROR', error: extractError(err, 'Hub list failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.KEY_LABEL_HUB_DETAIL,
-    async (_event, hubPostId: unknown): Promise<KeyLabelStoreResult<HubKeyLabelItem>> => {
-      try {
-        if (typeof hubPostId !== 'string' || !POST_ID_RE.test(hubPostId)) {
-          return { success: false, errorCode: 'NOT_FOUND', error: 'Invalid hub post id' }
-        }
-        const detail = await fetchKeyLabelDetail(hubPostId)
-        return { success: true, data: detail }
-      } catch (err) {
-        return { success: false, errorCode: 'IO_ERROR', error: extractError(err, 'Hub detail failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.KEY_LABEL_HUB_TIMESTAMPS,
-    async (_event, ids: unknown): Promise<KeyLabelStoreResult<HubKeyLabelTimestampsResponse>> => {
-      if (!Array.isArray(ids) || !ids.every((id) => typeof id === 'string' && POST_ID_RE.test(id))) {
-        return { success: false, errorCode: 'NOT_FOUND', error: 'ids must be an array of valid hub post ids' }
-      }
-      const unique = Array.from(new Set(ids as string[]))
-      if (unique.length === 0) return { success: true, data: { items: [] } }
-      try {
-        // Server caps each request at 100 ids; split larger inputs and
-        // run the chunks in parallel. Order is rebuilt from the input
-        // array so callers see input-order semantics regardless of
-        // chunking.
-        const chunks: string[][] = []
-        for (let i = 0; i < unique.length; i += HUB_KEY_LABEL_TIMESTAMPS_BATCH_LIMIT) {
-          chunks.push(unique.slice(i, i + HUB_KEY_LABEL_TIMESTAMPS_BATCH_LIMIT))
-        }
-        const responses = await Promise.all(chunks.map((chunk) => fetchKeyLabelTimestamps(chunk)))
-        const byId = new Map<string, HubKeyLabelTimestamp>()
-        for (const r of responses) {
-          for (const item of r.items) byId.set(item.id, item)
-        }
-        const items: HubKeyLabelTimestamp[] = []
-        for (const id of unique) {
-          const found = byId.get(id)
-          if (found) items.push(found)
-        }
-        return { success: true, data: { items } }
-      } catch (err) {
-        return { success: false, errorCode: 'IO_ERROR', error: extractError(err, 'Hub timestamps failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.KEY_LABEL_HUB_DOWNLOAD,
-    async (_event, hubPostId: unknown): Promise<KeyLabelStoreResult<KeyLabelMeta>> => {
-      try {
-        if (typeof hubPostId !== 'string' || !POST_ID_RE.test(hubPostId)) {
-          return { success: false, errorCode: 'NOT_FOUND', error: 'Invalid hub post id' }
-        }
-        const { body, uploaderName, hubUpdatedAt } = await fetchHubKeyLabelPayload(hubPostId)
-        // Use the Hub post id as the local id so the saved
-        // `keyboardLayout` can be matched against Hub later (e.g. the
-        // Missing Key Label dialog needs to look up the human name
-        // after the entry has been removed locally).
-        return await saveRecord({
-          id: hubPostId,
-          name: body.name,
-          hubPostId,
-          ...hubBodyToSaveRecordFields(body, uploaderName, hubUpdatedAt),
-        })
-      } catch (err) {
-        return { success: false, errorCode: 'IO_ERROR', error: extractError(err, 'Hub download failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.KEY_LABEL_HUB_UPLOAD,
-    async (_event, localId: unknown): Promise<KeyLabelStoreResult<KeyLabelMeta>> => {
-      if (typeof localId !== 'string') {
-        return { success: false, errorCode: 'NOT_FOUND', error: 'Invalid id' }
-      }
-      const record = await getRecord(localId)
-      if (!record.success || !record.data) {
-        return toKeyLabelLookupFailure(record)
-      }
-      const input = buildHubKeyLabelInput(record.data)
-      try {
-        const result = await withTokenRetry((jwt) => uploadKeyLabel(jwt, input))
-        // Carry the response's uploader_name and updated_at into the
-        // local meta so the modal immediately shows the Author and
-        // Updated columns and the Update / Remove buttons (gated by
-        // isMine = author === currentDisplayName) appear without
-        // waiting for a sync.
-        return setHubPostId(localId, result.id, result.uploader_name, result.updated_at)
-      } catch (err) {
-        if (err instanceof Hub409Error) {
-          return { success: false, errorCode: 'DUPLICATE_NAME', error: HUB_ERROR_KEY_LABEL_DUPLICATE }
-        }
-        return { success: false, errorCode: 'IO_ERROR', error: extractError(err, 'Hub upload failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.KEY_LABEL_HUB_UPDATE,
-    async (_event, localId: unknown): Promise<KeyLabelStoreResult<KeyLabelMeta>> => {
-      if (typeof localId !== 'string') {
-        return { success: false, errorCode: 'NOT_FOUND', error: 'Invalid id' }
-      }
-      const record = await getRecord(localId)
-      if (!record.success || !record.data) {
-        return toKeyLabelLookupFailure(record)
-      }
-      const hubPostId = record.data.meta.hubPostId
-      if (!hubPostId) {
-        return { success: false, errorCode: 'NOT_FOUND', error: 'Entry has no hub post' }
-      }
-      const input = buildHubKeyLabelInput(record.data)
-      try {
-        const result = await withTokenRetry((jwt) => updateKeyLabel(jwt, hubPostId, input))
-        // Persist the new Hub-side updated_at so the Updated column
-        // matches Hub's own display. Pass undefined uploaderName so
-        // setHubPostId leaves the existing value alone.
-        return setHubPostId(localId, hubPostId, undefined, result.updated_at)
-      } catch (err) {
-        if (err instanceof Hub409Error) {
-          return { success: false, errorCode: 'DUPLICATE_NAME', error: HUB_ERROR_KEY_LABEL_DUPLICATE }
-        }
-        return { success: false, errorCode: 'IO_ERROR', error: extractError(err, 'Hub update failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.KEY_LABEL_HUB_SYNC,
-    async (_event, localId: unknown): Promise<KeyLabelStoreResult<KeyLabelMeta>> => {
-      if (typeof localId !== 'string') {
-        return { success: false, errorCode: 'NOT_FOUND', error: 'Invalid id' }
-      }
-      const record = await getRecord(localId)
-      if (!record.success || !record.data) {
-        return toKeyLabelLookupFailure(record)
-      }
-      const hubPostId = record.data.meta.hubPostId
-      if (!hubPostId) {
-        return { success: false, errorCode: 'NOT_FOUND', error: 'Entry has no hub post' }
-      }
-      try {
-        const { body, uploaderName, hubUpdatedAt } = await fetchHubKeyLabelPayload(
-          hubPostId,
-          record.data.meta.uploaderName,
-          record.data.meta.hubUpdatedAt,
-        )
-        // Preserve the local id, name (drag/rename), and hubPostId; only
-        // refresh the payload (map / compositeLabels / keymapApplicable),
-        // uploaderName, and hubUpdatedAt.
-        return await saveRecord({
-          id: localId,
-          name: record.data.meta.name,
-          hubPostId,
-          ...hubBodyToSaveRecordFields(body, uploaderName, hubUpdatedAt),
-        })
-      } catch (err) {
-        return { success: false, errorCode: 'IO_ERROR', error: extractError(err, 'Hub sync failed') }
-      }
-    },
-  )
-
-  secureHandle(
-    IpcChannels.KEY_LABEL_HUB_DELETE,
-    async (_event, localId: unknown): Promise<KeyLabelStoreResult<void>> => {
-      if (typeof localId !== 'string') {
-        return { success: false, errorCode: 'NOT_FOUND', error: 'Invalid id' }
-      }
-      const record = await getRecord(localId)
-      if (!record.success || !record.data) return record as KeyLabelStoreResult<void>
-      const hubPostId = record.data.meta.hubPostId
-      if (!hubPostId) {
-        return { success: false, errorCode: 'NOT_FOUND', error: 'Entry has no hub post' }
-      }
-      try {
-        await withTokenRetry((jwt) => deleteKeyLabel(jwt, hubPostId))
-        const cleared = await setHubPostId(localId, null)
-        if (!cleared.success) return cleared as KeyLabelStoreResult<void>
-        return { success: true }
-      } catch (err) {
-        return { success: false, errorCode: 'IO_ERROR', error: extractError(err, 'Hub delete failed') }
-      }
-    },
-  )
+  registerHubPostHandlers()
+  registerHubFavoriteHandlers()
+  registerHubAnalyticsHandlers()
+  registerHubPrivateHandlers()
+  registerHubPackHandlers()
+  registerHubKeyLabelHandlers()
 }
+
+export { clearHubTokenCache } from './hub-ipc-token'
